+++ packages/frontend-next/src/components/report/ReportForm.tsx
@@ -1,14 +1,6 @@
 import { useNavigate } from '@tanstack/react-router';
 import { useTranslation } from 'react-i18next';
 
-import {
-  compareLineOrder,
-  type LineType,
-  resolveStationLineNames,
-  type Station,
-  useLines,
-  useStations,
-} from '@/api/transit';
 import { PageHeader } from '@/components/templates/PageHeader';
 import { LineBadge } from '@/components/transit/LineBadge';
 import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
@@ -35,24 +27,7 @@ function ClearSelectionButton({ onClick }: { onClick: () => void }) {
 
 function LinePicker() {
   const { t } = useTranslation(NAMESPACE);
-  const { lineName, setLineName, lineFilter, setLineFilter, stationId } = useReportSelection();
-  const { data: lines } = useLines();
-  const { data: stations } = useStations();
-
-  // One badge per line name (collapsing per-direction variants).
-  const allLines = () => {
-    const typeByName = new Map<string, LineType>();
-    for (const line of lines ?? []) {
-      if (!typeByName.has(line.name)) typeByName.set(line.name, line.type);
-    }
-    return [...typeByName].map(([name, type]) => ({ name, type })).sort(compareLineOrder);
-  };
-
-  const station = stationId ? stations?.[stationId] : undefined;
-  const stationLineNames = station ? new Set(resolveStationLineNames(station.lines, lines)) : null;
-  const visibleLines = allLines()
-    .filter((l) => lineFilter === 'all' || l.type === lineFilter)
-    .filter((l) => !stationLineNames || stationLineNames.has(l.name));
+  const { lineName, lineFilter, setLineFilter, selectLine, visibleLines } = useReportSelection();
 
   return (
     <section className="px-4">
@@ -84,7 +59,7 @@ function LinePicker() {
 
       {lineName && (
         <div className="mb-1 flex justify-end">
-          <ClearSelectionButton onClick={() => setLineName(null)} />
+          <ClearSelectionButton onClick={() => selectLine(null)} />
         </div>
       )}
 
@@ -96,14 +71,7 @@ function LinePicker() {
               key={line.name}
               type="button"
               aria-pressed={isSelected}
-              onClick={() => {
-                if (isSelected) {
-                  setLineName(null);
-                  return;
-                }
-                setLineName(line.name);
-                setLineFilter(line.type);
-              }}
+              onClick={() => selectLine(isSelected ? null : line.name)}
               className="rounded-sm outline-none focus-visible:ring-2 focus-visible:ring-white/50"
             >
               <LineBadge name={line.name} className={cn(isSelected && 'ring-2 ring-white')} />
@@ -117,45 +85,7 @@ function LinePicker() {
 
 function StationPicker() {
   const { t } = useTranslation(NAMESPACE);
-  const { lineName, lineFilter, stationId, setStationId } = useReportSelection();
-  const { data: stations } = useStations();
-  const { data: lines } = useLines();
-
-  const typeByName = new Map<string, LineType>();
-  for (const line of lines ?? []) typeByName.set(line.name, line.type);
-
-  const selectedStation = stationId ? stations?.[stationId] : undefined;
-
-  const stationsAlongLine = () => {
-    // Walk every variant of the selected line and emit stations in their stored order,
-    // deduplicating across direction variants.
-    const seen = new Set<string>();
-    const ordered: Station[] = [];
-    for (const line of lines ?? []) {
-      if (line.name !== lineName) continue;
-      for (const id of line.stations) {
-        if (seen.has(id)) continue;
-        seen.add(id);
-        const station = stations?.[id];
-        if (station) ordered.push(station);
-      }
-    }
-    return ordered;
-  };
-
-  const stationsByType = () =>
-    Object.values(stations ?? {})
-      .filter((station) => {
-        if (lineFilter === 'all') return true;
-        const names = resolveStationLineNames(station.lines, lines);
-        return names.some((name) => typeByName.get(name) === lineFilter);
-      })
-      .sort((a, b) => a.name.localeCompare(b.name));
-
-  let visibleStations: Station[];
-  if (selectedStation) visibleStations = [selectedStation];
-  else if (lineName) visibleStations = stationsAlongLine();
-  else visibleStations = stationsByType();
+  const { stationId, selectStation, visibleStations } = useReportSelection();
 
   return (
     <section className="mt-6 flex min-h-0 flex-1 flex-col px-4">
@@ -164,7 +94,7 @@ function StationPicker() {
           <h2 className="text-sm font-semibold">{t('station')}</h2>
           <p className="text-destructive text-[0.625rem] tracking-wide">{t('required')}</p>
         </div>
-        {stationId && <ClearSelectionButton onClick={() => setStationId(null)} />}
+        {stationId && <ClearSelectionButton onClick={() => selectStation(null)} />}
       </div>
       <ul className="min-h-0 flex-1 overflow-y-auto">
         {visibleStations.map((station) => {
@@ -174,7 +104,7 @@ function StationPicker() {
               <button
                 type="button"
                 aria-pressed={isSelected}
-                onClick={() => setStationId(isSelected ? null : station.id)}
+                onClick={() => selectStation(isSelected ? null : station.id)}
                 className={cn(
                   'hover:bg-muted focus-visible:bg-muted flex w-full items-center rounded-md px-2 py-2 text-left text-xs outline-none',
                   isSelected && 'ring-2 ring-white ring-inset',
